PMM-9637 create utils for labels
