+++ public/app/percona/integrated-alerting/components/Alerts/Alerts.constants.ts
@@ -1,15 +1,3 @@
 export const GET_ALERTS_CANCEL_TOKEN = 'getAlerts';
 export const ALERT_RULE_TEMPLATES_TABLE_ID = 'alerts';
 export const TOGGLE_ALERT_CANCEL_TOKEN = 'toggleAlert';
-export const PRIMARY_LABELS = [
-  'node_name',
-  'type',
-  'container_name',
-  'region',
-  'az',
-  'service_name',
-  'environment',
-  'cluster',
-  'replication_set',
-];
-export const HIDDEN_LABELS = ['alertname', 'job', 'rule_id', 'template_name', 'severity', 'ia'];
